aaaa

+++ local_goal_creator/include/local_goal_creator/local_goal_creator.h
@@ -20,7 +20,7 @@ private:
 
     int hz_;
     int index_step_;
-    int goal_index_;
+    double goal_index_;
     double taeget_distance_;
     bool is_path_ = false;
